+++ CHANGES.md
@@ -70,6 +70,12 @@ OpenSSL Releases
 
    *William McCormack*
 
+ * The B<openssl pkcs12 -export> command now accepts key and certificate(s)
+   in any order in the B<-in> file when B<-inkey> is not used.
+   Previously the private key had to appear before the certificates.
+
+   *John Claus*
+
 ### Changes between 3.6 and 4.0 [xx XXX xxxx]
 
  * Added support for RFC 8701 GREASE (Generate Random Extensions And Sustain
@@ -80,8 +86,6 @@ OpenSSL Releases
    `-grease` option to enable this.
 
    *William McCormack*
-
-### Changes between 3.6 and 4.0 [xx XXX xxxx]
 
  * Added `-expected-rpks` option to the `openssl s_client`
    and `openssl s_server` commands.  This makes it possible to specify